+++ android/src/main/kotlin/dev/steenbakker/mobile_scanner/MobileScanner.kt
@@ -13,15 +13,20 @@ import android.hardware.display.DisplayManager
 import android.net.Uri
 import android.os.Handler
 import android.os.Looper
+import android.util.Log
 import android.util.Size
 import android.view.Surface
 import androidx.annotation.VisibleForTesting
+import androidx.camera.camera2.Camera2Config
 import androidx.camera.core.Camera
 import androidx.camera.core.CameraSelector
+import androidx.camera.core.CameraXConfig
 import androidx.camera.core.ExperimentalGetImage
+import androidx.camera.core.ExperimentalLensFacing
 import androidx.camera.core.ImageAnalysis
 import androidx.camera.core.ImageProxy
 import androidx.camera.core.Preview
+import androidx.camera.core.SurfaceRequest
 import androidx.camera.core.TorchState
 import androidx.camera.core.resolutionselector.ResolutionSelector
 import androidx.camera.core.resolutionselector.ResolutionStrategy
@@ -37,13 +42,14 @@ import dev.steenbakker.mobile_scanner.objects.DetectionSpeed
 import dev.steenbakker.mobile_scanner.objects.MobileScannerErrorCodes
 import dev.steenbakker.mobile_scanner.objects.MobileScannerStartParameters
 import dev.steenbakker.mobile_scanner.utils.YuvToRgbConverter
+import dev.steenbakker.mobile_scanner.utils.serialize
 import io.flutter.view.TextureRegistry
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.coroutineScope
 import kotlinx.coroutines.launch
 import java.io.ByteArrayOutputStream
 import java.io.IOException
+import java.util.concurrent.Executors
 import kotlin.math.roundToInt
 
 class MobileScanner(
@@ -51,15 +57,20 @@ class MobileScanner(
     private val textureRegistry: TextureRegistry,
     private val mobileScannerCallback: MobileScannerCallback,
     private val mobileScannerErrorCallback: MobileScannerErrorCallback,
+    private val deviceOrientationListener: DeviceOrientationListener,
     private val barcodeScannerFactory: (options: BarcodeScannerOptions?) -> BarcodeScanner = ::defaultBarcodeScannerFactory,
 ) {
+
+    init {
+        configureCameraProcessProvider()
+    }
 
     /// Internal variables
     private var cameraProvider: ProcessCameraProvider? = null
     private var camera: Camera? = null
     private var cameraSelector: CameraSelector? = null
     private var preview: Preview? = null
-    private var textureEntry: TextureRegistry.SurfaceTextureEntry? = null
+    private var surfaceProducer: TextureRegistry.SurfaceProducer? = null
     private var scanner: BarcodeScanner? = null
     private var lastScanned: List<String?>? = null
     private var scannerTimeout = false
@@ -74,6 +85,20 @@ class MobileScanner(
     private var isPaused = false
 
     companion object {
+        // Configure the `ProcessCameraProvider` to only log errors.
+        // This prevents the informational log spam from CameraX.
+        private fun configureCameraProcessProvider() {
+            try {
+                val config = CameraXConfig.Builder.fromConfig(Camera2Config.defaultConfig()).apply {
+                    setMinimumLoggingLevel(Log.ERROR)
+                }
+                ProcessCameraProvider.configureInstance(config.build())
+            } catch (_: IllegalStateException) {
+                // The ProcessCameraProvider was already configured.
+                // Do nothing.
+            }
+        }
+
         /**
          * Create a barcode scanner from the given options.
          */
@@ -140,12 +165,12 @@ class MobileScanner(
                 val portrait = (camera?.cameraInfo?.sensorRotationDegrees ?: 0) % 180 == 0
 
                 if (!returnImage) {
-                    imageProxy.close()
                     mobileScannerCallback(
                         barcodeMap,
                         null,
                         if (portrait) inputImage.width else inputImage.height,
                         if (portrait) inputImage.height else inputImage.width)
+                    imageProxy.close()
                     return@addOnSuccessListener
                 }
 
@@ -163,15 +188,16 @@ class MobileScanner(
                     val bmWidth = bmResult.width
                     val bmHeight = bmResult.height
 
-                    bmResult.recycle()
-                    imageProxy.close()
-
                     mobileScannerCallback(
                         barcodeMap,
                         byteArray,
                         bmWidth,
                         bmHeight
                     )
+
+                    bmResult.recycle()
+                    imageProxy.close()
+                    imageFormat.release()
                 }
 
             }.addOnFailureListener { e ->
@@ -186,6 +212,81 @@ class MobileScanner(
             Handler(Looper.getMainLooper()).postDelayed({
                 scannerTimeout = false
             }, detectionTimeout)
+        }
+    }
+
+    /**
+     * Create a {@link Preview.SurfaceProvider} that specifies how to provide a {@link Surface} to a
+     * {@code Preview}.
+     */
+    @VisibleForTesting
+    fun createSurfaceProvider(surfaceProducer: TextureRegistry.SurfaceProducer): Preview.SurfaceProvider {
+        return Preview.SurfaceProvider {
+            request: SurfaceRequest ->
+            run {
+                // Set the callback for the surfaceProducer to invalidate Surfaces that it produces
+                // when they get destroyed.
+                surfaceProducer.setCallback(
+                    object : TextureRegistry.SurfaceProducer.Callback {
+                        override fun onSurfaceAvailable() {
+                            // Do nothing. The Preview.SurfaceProvider will handle this
+                            // whenever a new Surface is needed.
+                        }
+
+                        // TODO: replace with "onSurfaceCleanup" when available in Flutter 3.28 or later
+                        // See https://github.com/flutter/flutter/pull/160937
+                        override fun onSurfaceDestroyed() {
+                            // Invalidate the SurfaceRequest so that CameraX knows to to make a new request
+                            // for a surface.
+                            request.invalidate()
+                        }
+                    }
+                )
+
+                // Provide the surface.
+                surfaceProducer.setSize(request.resolution.width, request.resolution.height)
+
+                val surface: Surface = surfaceProducer.surface
+
+                // The single thread executor is only used to invoke the result callback.
+                // Thus it is safe to use a new executor,
+                // instead of reusing the executor that is passed to the camera process provider.
+                request.provideSurface(surface, Executors.newSingleThreadExecutor()) {
+                    // Handle the result of the request for a surface.
+                    // See: https://developer.android.com/reference/androidx/camera/core/SurfaceRequest.Result
+
+                    // Always attempt a release.
+                    surface.release()
+
+                    val resultCode: Int = it.resultCode
+
+                    when(resultCode) {
+                        SurfaceRequest.Result.RESULT_REQUEST_CANCELLED,
+                        SurfaceRequest.Result.RESULT_WILL_NOT_PROVIDE_SURFACE,
+                        SurfaceRequest.Result.RESULT_SURFACE_ALREADY_PROVIDED,
+                        SurfaceRequest.Result.RESULT_SURFACE_USED_SUCCESSFULLY -> {
+                            // Only need to release, do nothing.
+                        }
+                        SurfaceRequest.Result.RESULT_INVALID_SURFACE -> {
+                            // The surface was invalid, so it is not clear how to recover from this.
+                        }
+                        else -> {
+                            // Fallthrough, in case any result codes are added later.
+                        }
+                    }
+                }
+            }
+        }
+    }
+
+    @ExperimentalLensFacing
+    private fun getCameraLensFacing(camera: Camera?): Int? {
+        return when(camera?.cameraInfo?.lensFacing) {
+            CameraSelector.LENS_FACING_BACK -> 1
+            CameraSelector.LENS_FACING_FRONT -> 0
+            CameraSelector.LENS_FACING_EXTERNAL -> 2
+            CameraSelector.LENS_FACING_UNKNOWN -> null
+            else -> null
         }
     }
 
@@ -229,6 +330,7 @@ class MobileScanner(
     /**
      * Start barcode scanning by initializing the camera and barcode scanner.
      */
+    @ExperimentalLensFacing
     @ExperimentalGetImage
     fun start(
         barcodeScannerOptions: BarcodeScannerOptions?,
@@ -249,19 +351,24 @@ class MobileScanner(
         this.returnImage = returnImage
         this.invertImage = invertImage
 
-        if (camera?.cameraInfo != null && preview != null && textureEntry != null && !isPaused) {
+        if (camera?.cameraInfo != null && preview != null && surfaceProducer != null && !isPaused) {
 
-           // TODO: resume here for seamless transition
+// TODO: resume here for seamless transition
 //            if (isPaused) {
 //                resumeCamera()
+//                val cameraDirection = getCameraLensFacing(camera)
 //                mobileScannerStartedCallback(
-//                    MobileScannerStartParameters(
-//                        if (portrait) width else height,
-//                        if (portrait) height else width,
-//                        currentTorchState,
-//                        textureEntry!!.id(),
-//                        numberOfCameras ?: 0
-//                    )
+//                  MobileScannerStartParameters(
+//                    if (portrait) width else height,
+//                    if (portrait) height else width,
+//                    deviceOrientationListener.getUIOrientation().serialize(),
+//                    sensorRotationDegrees,
+//                    surfaceProducer!!.handlesCropAndRotation(),
+//                    currentTorchState,
+//                    surfaceProducer!!.id(),
+//                    numberOfCameras ?: 0,
+//                    cameraDirection
+//                  )
 //                )
 //                return
 //            }
@@ -287,23 +394,10 @@ class MobileScanner(
             }
 
             cameraProvider?.unbindAll()
-            textureEntry = textureEntry ?: textureRegistry.createSurfaceTexture()
+            surfaceProducer = surfaceProducer ?: textureRegistry.createSurfaceProducer()
+            val surfaceProvider: Preview.SurfaceProvider = createSurfaceProvider(surfaceProducer!!)
 
             // Preview
-            val surfaceProvider = Preview.SurfaceProvider { request ->
-                if (isStopped()) {
-                    return@SurfaceProvider
-                }
-
-                val texture = textureEntry!!.surfaceTexture()
-                texture.setDefaultBufferSize(
-                    request.resolution.width,
-                    request.resolution.height
-                )
-
-                val surface = Surface(texture)
-                request.provideSurface(surface, executor) { }
-            }
 
             // Build the preview to be shown on the Flutter texture
             val previewBuilder = Preview.Builder()
@@ -384,7 +478,9 @@ class MobileScanner(
             val resolution = analysis.resolutionInfo!!.resolution
             val width = resolution.width.toDouble()
             val height = resolution.height.toDouble()
-            val portrait = (camera?.cameraInfo?.sensorRotationDegrees ?: 0) % 180 == 0
+            val sensorRotationDegrees = camera?.cameraInfo?.sensorRotationDegrees ?: 0
+            val portrait = sensorRotationDegrees % 180 == 0
+            val cameraDirection = getCameraLensFacing(camera)
 
             // Start with 'unavailable' torch state.
             var currentTorchState: Int = -1
@@ -397,13 +493,19 @@ class MobileScanner(
                 currentTorchState = it.torchState.value ?: -1
             }
 
+            deviceOrientationListener.start()
+
             mobileScannerStartedCallback(
                 MobileScannerStartParameters(
                     if (portrait) width else height,
                     if (portrait) height else width,
+                    deviceOrientationListener.getUIOrientation().serialize(),
+                    sensorRotationDegrees,
+                    surfaceProducer!!.handlesCropAndRotation(),
                     currentTorchState,
-                    textureEntry!!.id(),
-                    numberOfCameras ?: 0
+                    surfaceProducer!!.id(),
+                    numberOfCameras ?: 0,
+                    cameraDirection,
                 )
             )
         }, executor)
@@ -420,6 +522,7 @@ class MobileScanner(
             throw AlreadyStopped()
         }
 
+        deviceOrientationListener.stop()
         pauseCamera()
     }
 
@@ -431,6 +534,7 @@ class MobileScanner(
             throw AlreadyStopped()
         }
 
+        deviceOrientationListener.stop()
         releaseCamera()
     }
 
@@ -469,8 +573,9 @@ class MobileScanner(
         // The camera will be closed when the last use case is unbound.
         cameraProvider?.unbindAll()
 
-        textureEntry?.release()
-        textureEntry = null
+        // Release the surface for the preview.
+        surfaceProducer?.release()
+        surfaceProducer = null
 
         // Release the scanner.
         scanner?.close()
@@ -533,7 +638,7 @@ class MobileScanner(
         }
         val paint = Paint().apply { colorFilter = ColorMatrixColorFilter(colorMatrix) }
 
-        val invertedBitmap = Bitmap.createBitmap(bitmap.width, bitmap.height, bitmap.config)
+        val invertedBitmap = Bitmap.createBitmap(bitmap.width, bitmap.height, bitmap.config!!)
         val canvas = Canvas(invertedBitmap)
         canvas.drawBitmap(bitmap, 0f, 0f, paint)