+++ lib/src/method_channel/mobile_scanner_method_channel.dart
@@ -4,15 +4,19 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
 import 'package:mobile_scanner/src/enums/barcode_format.dart';
+import 'package:mobile_scanner/src/enums/camera_facing.dart';
 import 'package:mobile_scanner/src/enums/mobile_scanner_authorization_state.dart';
 import 'package:mobile_scanner/src/enums/mobile_scanner_error_code.dart';
 import 'package:mobile_scanner/src/enums/torch_state.dart';
+import 'package:mobile_scanner/src/method_channel/android_surface_producer_delegate.dart';
+import 'package:mobile_scanner/src/method_channel/rotated_preview.dart';
 import 'package:mobile_scanner/src/mobile_scanner_exception.dart';
 import 'package:mobile_scanner/src/mobile_scanner_platform_interface.dart';
 import 'package:mobile_scanner/src/mobile_scanner_view_attributes.dart';
 import 'package:mobile_scanner/src/objects/barcode.dart';
 import 'package:mobile_scanner/src/objects/barcode_capture.dart';
 import 'package:mobile_scanner/src/objects/start_options.dart';
+import 'package:mobile_scanner/src/utils/parse_device_orientation_extension.dart';
 
 /// An implementation of [MobileScannerPlatform] that uses method channels.
 class MethodChannelMobileScanner extends MobileScannerPlatform {
@@ -35,13 +39,31 @@ class MethodChannelMobileScanner extends MobileScannerPlatform {
     'dev.steenbakker.mobile_scanner/scanner/method',
   );
 
+  /// The event channel that sends back device orientation change events.
+  @visibleForTesting
+  final deviceOrientationEventChannel = const EventChannel(
+    'dev.steenbakker.mobile_scanner/scanner/deviceOrientation',
+  );
+
   /// The event channel that sends back scanned barcode events.
   @visibleForTesting
   final eventChannel = const EventChannel(
     'dev.steenbakker.mobile_scanner/scanner/event',
   );
 
+  Stream<DeviceOrientation>? _deviceOrientationStream;
   Stream<Map<Object?, Object?>>? _eventsStream;
+
+  /// Get the event stream of device orientation change events
+  /// that come from the [deviceOrientationEventChannel].
+  Stream<DeviceOrientation> get deviceOrientationChangedStream {
+    _deviceOrientationStream ??= deviceOrientationEventChannel
+        .receiveBroadcastStream()
+        .cast<String>()
+        .map((String orientation) => orientation.parseDeviceOrientation());
+
+    return _deviceOrientationStream!;
+  }
 
   /// Get the event stream of barcode events that come from the [eventChannel].
   Stream<Map<Object?, Object?>> get eventsStream {
@@ -51,6 +73,10 @@ class MethodChannelMobileScanner extends MobileScannerPlatform {
     return _eventsStream!;
   }
 
+  /// The delegate that handles texture rotation corrections on Android.
+  AndroidSurfaceProducerDelegate? _surfaceProducerDelegate;
+
+  /// The identifier of the current texture.
   int? _textureId;
   bool _pausing = false;
 
@@ -212,7 +238,25 @@ class MethodChannelMobileScanner extends MobileScannerPlatform {
       return const SizedBox();
     }
 
-    return Texture(textureId: _textureId!);
+    final Widget texture = Texture(textureId: _textureId!);
+
+    // If the preview needs manual orientation corrections,
+    // correct the preview orientation based on the currently reported device orientation.
+    // On Android, the underlying device orientation stream will emit the current orientation
+    // when the first listener is attached.
+    if (_surfaceProducerDelegate
+        case final AndroidSurfaceProducerDelegate delegate
+        when !delegate.handlesCropAndRotation) {
+      return RotatedPreview.fromCameraDirection(
+        delegate.cameraFacingDirection,
+        deviceOrientationStream: deviceOrientationChangedStream,
+        initialDeviceOrientation: delegate.initialDeviceOrientation,
+        sensorOrientationDegrees: delegate.sensorOrientationDegrees,
+        child: texture,
+      );
+    }
+
+    return texture;
   }
 
   @override
@@ -276,7 +320,18 @@ class MethodChannelMobileScanner extends MobileScannerPlatform {
       );
     }
 
+    final CameraFacing cameraDirection =
+        CameraFacing.fromRawValue(startResult['cameraDirection'] as int?);
+
     _textureId = textureId;
+
+    if (defaultTargetPlatform == TargetPlatform.android) {
+      _surfaceProducerDelegate =
+          AndroidSurfaceProducerDelegate.fromConfiguration(
+        startResult,
+        cameraDirection,
+      );
+    }
 
     final int? numberOfCameras = startResult['numberOfCameras'] as int?;
     final TorchState currentTorchState = TorchState.fromRawValue(
@@ -295,6 +350,7 @@ class MethodChannelMobileScanner extends MobileScannerPlatform {
     _pausing = false;
 
     return MobileScannerViewAttributes(
+      cameraDirection: cameraDirection,
       currentTorchMode: currentTorchState,
       numberOfCameras: numberOfCameras,
       size: size,
@@ -309,6 +365,7 @@ class MethodChannelMobileScanner extends MobileScannerPlatform {
 
     _textureId = null;
     _pausing = false;
+    _surfaceProducerDelegate = null;
 
     await methodChannel.invokeMethod<void>('stop');
   }